Reecriture d'une partie du code pour qu'elle soit plus lisible et plus simple et ajout d'un exemple

+++ src/main/java/fr/hexagone/front/Coordinate.java
@@ -2,10 +2,10 @@ package fr.hexagone.front;
 
 public class Coordinate {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Coordinate(int x, int y){
+    public Coordinate(double x, double y){
         this.x = x;
         this.y = y;
     }
